requirements.txt
python-telegram-bot==13.2
mysql-connector-python==8.0.23